Trying to abstract the navigate board method
I used lambdas to navigate the board, but i need to reuse the same lambda. Refactoring is proving difficult as I'm essentially recreating nested for loops in a lambda, which isn;t the best thing to do.

The current lambda boxes integers into a collection each time, and depending on direction, reverses each  the collections. I think this is the cause of performance issues. I'll try to fix this by using pre built lists.