Functionally complete
Recreating 2048 has been fun. It's been one of my favourite games but always baffled me about how the logic works. Thus, I decided to create it myself and learn the logic behind it.
I've got the application functionally complete, but the UI is needs improvement, but I've learned all I wanted to from this and so not progressing any further.